Move core schema generation logic for path types inside the `GenerateSchema` class (#10846)
